Crate binomial_iter

This crate provides tools to iteratively compute the binomial coefficient.

Structs

BinomialIter
Provides methods to calculate the binomial coefficient for the next higher/lower n/k.

DecKIter
/// An iterator which wraps a BinomialIter and returns the result of it's /// dec_k method when next is called.

DecNIter
/// An iterator which wraps a BinomialIter and returns the result of it's /// dec_n method when next is called.

IncKIter
/// An iterator which wraps a BinomialIter and returns the result of it's /// inc_k method when next is called.

IncNIter
/// An iterator which wraps a BinomialIter and returns the result of it's /// inc_n method when next is called.