Struct binomial_iter::BinomialIter

pub struct BinomialIter {
    // some fields omitted
}

Provides methods to calculate the binomial coefficient for the next higher/lower n/k.

Methods

impl BinomialIter

fn new(n: u32, k: u32) -> BinomialIter

Create a new BinomialIter. This will calculate the binomial coefficient once using the recursive definition.

Panics

If k > n

fn n(&self) -> u32

Access the current value of n.

fn k(&self) -> u32

Access the current value of k.

fn binom(&self) -> u32

Access the current value of n choose k.

fn inc_n(&mut self) -> Option<(u32, u32)>

Increase n by one and update the internal state accordingly.

Returns None when calculating n + 1 choose k would overflow, otherwise Some((n + 1, binom)).

fn dec_n(&mut self) -> Option<(u32, u32)>

Decrease n by one and update the internal state accordingly.

Returns None when calculating n - 1 choose k would overflow or n - 1 < k, otherwise Some((n - 1, binom)).

fn inc_k(&mut self) -> Option<(u32, u32)>

Increase k by one and update the internal state accordingly.

Returns None when calculating n choose k + 1 would overflow or n < k + 1, otherwise Some((k + 1, binom)).

fn dec_k(&mut self) -> Option<(u32, u32)>

Decrease k by one and update the internal state accordingly.

Returns None when calculating n choose k - 1 would overflow or k - 1 < 0, otherwise Some((k - 1, binom))

impl BinomialIter

fn iter_inc_n(self) -> IncNIter

/// Returns an iterator which wraps this BinomialIter, returns the current /// value of n and binom on the first call to next and the result of /// calling inc_n on the underlying BinominalIter for subsequent calls /// to next.

impl BinomialIter

fn iter_dec_n(self) -> DecNIter

/// Returns an iterator which wraps this BinomialIter, returns the current /// value of n and binom on the first call to next and the result of /// calling dec_n on the underlying BinominalIter for subsequent calls /// to next.

impl BinomialIter

fn iter_inc_k(self) -> IncKIter

/// Returns an iterator which wraps this BinomialIter, returns the current /// value of k and binom on the first call to next and the result of /// calling inc_k on the underlying BinominalIter for subsequent calls /// to next.

impl BinomialIter

fn iter_dec_k(self) -> DecKIter

/// Returns an iterator which wraps this BinomialIter, returns the current /// value of k and binom on the first call to next and the result of /// calling dec_k on the underlying BinominalIter for subsequent calls /// to next.